+++ internal/origins/drivers/opkg/parser.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"path/filepath"
 	"qp/internal/consts"
-	"qp/internal/origins/formats/debstyle"
+	"qp/internal/origins/shared/debstyle"
 	"qp/internal/origins/worker"
 	"qp/internal/pkgdata"
 	"strconv"
